requirements.txt
BeautifulSoup==3.2.1
Django==1.5.1
dj-database-url==0.3.0
dj-static==0.0.6
django-annoying==0.8.0
django-oauth-tokens==0.3.12
django-picklefield==0.3.1
django-taggit==0.12.2
django-toolbelt==0.0.1
django-twitter-api==0.1.4
freeze==0.8.0
gunicorn==19.1.1
oauthlib==0.6.3
psycopg2==2.5.4
python-dateutil==2.2
pytz==2014.7
requests==2.1.0
requests-oauthlib==0.4.0
six==1.8.0
static3==0.5.1
tweepy==2.3.0
twython==3.1.2
tyoi.OAuth2==0.2.1
wsgiref==0.1.2
